requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
bson
requests